[CuongTQ] fix API bugs

+++ SourceCode/WebServer/src/main/java/com/fpt/edu/controller/BookController.java
@@ -24,8 +24,12 @@ public class BookController extends BaseController {
 
     @ApiOperation(value = "Get a book by its id", response = String.class)
     @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = Constant.APPLICATION_JSON)
-    public ResponseEntity<String> getBookById(@PathVariable Long id) throws JsonProcessingException {
+    public ResponseEntity<String> getBookById(@PathVariable Long id) throws JsonProcessingException, EntityNotFoundException {
         Book book = bookServices.getBookById(id);
+        if (book == null) {
+            throw new EntityNotFoundException("Book id: " + id + " not found");
+        }
+
         JSONObject jsonObject = utils.convertObjectToJSONObject(book);
         return new ResponseEntity<>(jsonObject.toString(), HttpStatus.OK);
     }
@@ -40,8 +44,18 @@ public class BookController extends BaseController {
 
     @ApiOperation(value = "Update a book", response = String.class)
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT, produces = Constant.APPLICATION_JSON)
-    public ResponseEntity<String> updateBook(@PathVariable Long id, @RequestBody String body) throws IOException, EntityNotFoundException {
-        bookServices.getBookById(id);
+    public ResponseEntity<String> updateBook(@PathVariable Long id, @RequestBody String body) throws IOException, EntityNotFoundException, EntityIdMismatchException {
+        JSONObject bodyObject = new JSONObject(body);
+        Long bookId = bodyObject.getLong("id");
+
+        if (bookId != id) {
+            throw new EntityIdMismatchException("Book ID: " + id + " and " + bookId + " does not match");
+        }
+
+        Book existedBook = bookServices.getBookById(id);
+        if (existedBook == null) {
+            throw new EntityNotFoundException("Book id: " + id + " not found");
+        }
 
         Book bookResult = bookServices.updateBook(body);
 
@@ -51,8 +65,11 @@ public class BookController extends BaseController {
 
     @ApiOperation(value = "Delete a book", response = String.class)
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = Constant.APPLICATION_JSON)
-    public ResponseEntity<String> deleteBook(@PathVariable Long id) {
-        bookServices.getBookById(id);
+    public ResponseEntity<String> deleteBook(@PathVariable Long id) throws EntityNotFoundException {
+        Book existedBook = bookServices.getBookById(id);
+        if (existedBook == null) {
+            throw new EntityNotFoundException("Book id: " + id + " not found");
+        }
 
         bookServices.deleteBook(id);
         JSONObject jsonObject = new JSONObject();
